+++ tests/unit/odt/api/test_sbds.py
@@ -154,7 +154,7 @@ class TestSBDefinitionAPI:
         """
         mock_validate.return_value = [
             ValidationIssue(
-                message="Maximum elevation (44.74 degrees) is less than the limit (45.0 degrees) ",
+                message="Maximum elevation (44.74 degrees) is less than the limit (45.0 degrees)",
                 field="$.targets.0",
                 level=ValidationIssueType.ERROR,
             )
@@ -166,17 +166,9 @@ class TestSBDefinitionAPI:
         )
 
         assert response.json() == {
-            "detail": {
-                "valid": False,
-                "issues": [
-                    {
-                        "field": "$.targets.0",
-                        "level": "error",
-                        "message": "Maximum elevation (44.74 degrees) is less "
-                        "than the limit (45.0 degrees) ",
-                    }
-                ],
-            }
+            "detail": "SBDefinition validation failed with issues "
+            "['$.targets.0: Maximum elevation (44.74 degrees) "
+            "is less than the limit (45.0 degrees)']"
         }
         assert response.status_code == HTTPStatus.BAD_REQUEST
 
@@ -258,10 +250,7 @@ class TestSBDefinitionAPI:
 
         assert response.status_code == HTTPStatus.BAD_REQUEST
         assert response.json() == {
-            "detail": {
-                "issues": [{"field": "$.targets.0", "level": "error", "message": "foo"}],
-                "valid": False,
-            }
+            "detail": "SBDefinition validation failed with issues ['$.targets.0: foo']"
         }
 
     @mock.patch("ska_oso_services.odt.api.sbds.validate_sbdefinition")