+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Unreleased
 **********
 * Adds `/calibrators` endpoint for PHT that returns the appropriate calibrators for a given strategy
+* [BUGFIX] Fixes bugs in `generateBasicCommissioningSBDefinition` where reference frame was not passed into function and correct defaults not set
 
 16.2.0
 **********